+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.4.0
+- Improved default paths resolving
+- Added base dir to output options
+- Added optional include out path support
+- Improve logs format
+
 # 1.3.3
 - `new` command replaced with `generate`